pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-adapters"
version = "0.1.0"
description = "A2A Protocol Adapter SDK for integrating various agent frameworks"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "HYBRO AI", email = "info@hybro.ai"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "a2a-sdk>=0.1.0",
    "uvicorn>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
n8n = []
crewai = ["crewai>=0.1.0"]
langchain = ["langchain>=0.1.0", "langchain-core>=0.1.0"]
langgraph = ["langgraph>=0.0.1"]
all = [
    "crewai>=0.1.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/hybro-ai/a2a-adapters"
Documentation = "https://github.com/hybro-ai/a2a-adapters#readme"
Repository = "https://github.com/hybro-ai/a2a-adapters"

[tool.setuptools]
packages = ["a2a_adapters", "a2a_adapters.integrations"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
